ai-engine/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Deep Learning
tensorflow==2.15.0
torch==2.1.2
torchvision==0.16.2

# Computer Vision
opencv-python==4.9.0.80
mediapipe==0.10.9
Pillow==10.2.0

# Scientific Computing
numpy==1.26.3
scipy==1.12.0
scikit-learn==1.4.0

# Explainable AI
lime==0.2.0.1
shap==0.44.1

# Utilities
python-jose[cryptography]==3.3.0
aiofiles==23.2.1
httpx==0.26.0

# Image/Video Processing
imageio==2.33.1
imageio-ffmpeg==0.4.9

# Logging & Monitoring
loguru==0.7.2
